+++ scheduler/client.go
@@ -147,40 +147,16 @@ func (c *Client) GetState() (State, error) {
 
 // TriggerDagRun schedules new DAG run.
 func (c *Client) TriggerDagRun(in api.DagRunTriggerInput) error {
-	jsonInput, jErr := json.Marshal(in)
-	if jErr != nil {
-		return fmt.Errorf("cannot serialize DagRunTriggerInput: %s",
-			jErr.Error())
-	}
-	req, rErr := http.NewRequest(
-		"POST", c.triggerDagRunUrl(), bytes.NewBuffer(jsonInput),
+	return httpPost[api.DagRunTriggerInput](
+		c.httpClient, c.triggerDagRunUrl(), in,
 	)
-	if rErr != nil {
-		return fmt.Errorf("cannot create POST request: %s", rErr.Error())
-	}
-	req.Header.Set("Content-Type", "application/json")
-	defer req.Body.Close()
+}
 
-	resp, err := c.httpClient.Do(req)
-	if err != nil {
-		return fmt.Errorf("cannot perform POST request: %s", err.Error())
-	}
-	if resp.StatusCode != http.StatusOK {
-		defer resp.Body.Close()
-		body, rErr := io.ReadAll(resp.Body)
-		if rErr != nil {
-			msg := fmt.Sprintf(
-				"unexpected status code in TriggerDagRun request: %d, "+
-					"and cannot read the response body: %s",
-				resp.StatusCode, rErr.Error(),
-			)
-			return errors.New(msg)
-		}
-		return fmt.Errorf("unexpected status code in TriggerDagRun request: %d. Body: %s",
-			resp.StatusCode, string(body))
-	}
-
-	return nil
+// RestartDagRun restarts given DAG run.
+func (c *Client) RestartDagRun(in api.DagRunRestartInput) error {
+	return httpPost[api.DagRunRestartInput](
+		c.httpClient, c.restartDagRunUrl(), in,
+	)
 }
 
 // UIDagrunStats returns the current statistics on DAG runs, its tasks and
@@ -295,6 +271,11 @@ func (c *Client) triggerDagRunUrl() string {
 	return fmt.Sprintf("%s%s", c.schedulerUrl, suffix)
 }
 
+func (c *Client) restartDagRunUrl() string {
+	suffix := c.routes[api.EndpointDagRunRestart].UrlSuffix
+	return fmt.Sprintf("%s%s", c.schedulerUrl, suffix)
+}
+
 func (c *Client) getUpdateTaskStatusUrl() string {
 	suffix := c.routes[api.EndpointDagTaskUpdate].UrlSuffix
 	return fmt.Sprintf("%s%s", c.schedulerUrl, suffix)
@@ -319,34 +300,4 @@ func (c *Client) uiDagrunTaskDetailsUrl(runId int, taskId string, retry int) str
 	suffix := c.routes[api.EndpointUiDagrunTaskDetails].UrlSuffix
 	return fmt.Sprintf("%s%s/%d/%s/%d", c.schedulerUrl, suffix, runId,
 		taskId, retry)
-}
-
-// Generic HTTP GET request with resp body deserialization from JSON into given
-// type.
-func httpGetJSON[T any](client *http.Client, url string) (*T, int, error) {
-	resp, err := client.Get(url)
-	if err != nil {
-		return nil, http.StatusBadRequest,
-			fmt.Errorf("failed to perform GET request: %w", err)
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		return nil, resp.StatusCode,
-			fmt.Errorf("received non-200 response: %s", resp.Status)
-	}
-
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return nil, resp.StatusCode,
-			fmt.Errorf("failed to read response body: %w", err)
-	}
-
-	var result T
-	if err := json.Unmarshal(body, &result); err != nil {
-		return nil, resp.StatusCode,
-			fmt.Errorf("failed to unmarshal JSON: %w", err)
-	}
-
-	return &result, resp.StatusCode, nil
 }